Move attribute out of prefab.py for generation.
