+++ src/prefab_classes/live/prefab.py
@@ -51,7 +51,7 @@ from functools import partial
 # noinspection PyUnresolvedReferences
 # from typing import dataclass_transform
 
-
+from ._attribute_class import Attribute
 from ..constants import FIELDS_ATTRIBUTE, COMPILED_FLAG, CLASSVAR_NAME
 from ..exceptions import PrefabError, LivePrefabError, CompiledPrefabError
 from .default_sentinels import _NOTHING
@@ -62,74 +62,6 @@ from .method_generators import (
     iter_maker,
     prefab_init_maker,
 )
-
-
-class Attribute:
-    __slots__ = (
-        "default",
-        "default_factory",
-        "converter",
-        "init",
-        "repr",
-        "kw_only",
-    )
-
-    def __init__(
-        self,
-        *,
-        default=_NOTHING,
-        default_factory=_NOTHING,
-        converter=None,
-        init=True,
-        repr=True,
-        kw_only=False,
-    ):
-        """
-        Initialize an Attribute
-
-        :param default: Default value for this attribute
-        :param default_factory: No argument callable to give a default value
-                                (for otherwise mutable defaults)
-        :param converter: prefab.attr = x -> prefab.attr = converter(x)
-        :param init: Include this attribute in the __init__ parameters
-        :param repr: Include this attribute in the class __repr__
-        :param kw_only: Make this argument keyword only in init
-        """
-
-        if not init and default is _NOTHING and default_factory is _NOTHING:
-            raise LivePrefabError(
-                "Must provide a default value/factory "
-                "if the attribute is not in init."
-            )
-
-        if kw_only and not init:
-            raise LivePrefabError(
-                "Attribute cannot be keyword only if it is not in init."
-            )
-
-        if default is not _NOTHING and default_factory is not _NOTHING:
-            raise LivePrefabError(
-                "Cannot define both a default value and a default factory."
-            )
-
-        self.default = default
-        self.default_factory = default_factory
-        self.converter = converter
-        self.init = init
-        self.repr = repr
-        self.kw_only = kw_only
-
-    def __repr__(self):
-        return (
-            f"Attribute("
-            f"default={self.default!r}, "
-            f"default_factory={self.default_factory!r}, "
-            f"converter={self.converter!r}, "
-            f"init={self.init!r}, "
-            f"repr={self.repr!r}, "
-            f"kw_only={self.kw_only!r}"
-            f")"
-        )
 
 
 def attribute(
@@ -166,7 +98,9 @@ def attribute(
 
 
 # @dataclass_transform(field_specifiers=(attribute, Attribute))
-def _make_prefab(cls: type, *, init=True, repr=True, eq=True, iter=False, match_args=True):
+def _make_prefab(
+    cls: type, *, init=True, repr=True, eq=True, iter=False, match_args=True
+):
     """
     Generate boilerplate code for dunder methods in a class.
 
@@ -188,12 +122,15 @@ def _make_prefab(cls: type, *, init=True, repr=True, eq=True, iter=False, match_
     # Eliminate ClassVars - don't want to import typing if
     # it's not already imported
     if annotations:
-        _typing = sys.modules.get('typing')
+        _typing = sys.modules.get("typing")
         if _typing:
             new_annotations = {}
             for key, value in annotations.items():
                 # Actual class used as annotation
-                if value is _typing.ClassVar or getattr(value, '__origin__', None) is _typing.ClassVar:
+                if (
+                    value is _typing.ClassVar
+                    or getattr(value, "__origin__", None) is _typing.ClassVar
+                ):
                     continue
                 # String used as annotation
                 elif isinstance(value, str) and CLASSVAR_NAME in value:
@@ -347,4 +284,6 @@ def prefab(
         else:
             # Create Live Version
             setattr(cls, COMPILED_FLAG, False)
-            return _make_prefab(cls, init=init, repr=repr, eq=eq, iter=iter, match_args=match_args)
+            return _make_prefab(
+                cls, init=init, repr=repr, eq=eq, iter=iter, match_args=match_args
+            )
